# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/v1/menus:
    get:
      tags:
        - Menus
      summary: List Menus
      description: Get a paginated list of Menus with optional filtering and sorting
      operationId: listMenus
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MenuList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Menus
      summary: Create Menu
      description: Create a new Menu
      operationId: createMenu
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMenuRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Menu'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/page_blocks:
    get:
      tags:
        - PageBlocks
      summary: List PageBlocks
      description: Get a paginated list of PageBlocks with optional filtering and sorting
      operationId: listPageBlocks
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageBlockList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - PageBlocks
      summary: Create PageBlock
      description: Create a new PageBlock
      operationId: createPageBlock
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePageBlockRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageBlock'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
...
    Menu:
      type: object
      required:
        - id
        - website_id
        - name
        - new_window
        - sequence
        - visibility
        - required_member_roles
        - is_mega_menu
        - metadata
      properties:
        id:
          type: string
          format: uuid
        website_id:
          type: string
          format: uuid
        parent_id:
          nullable: true
          type: string
          format: uuid
        name:
          type: string
          maxLength: 120
        page_id:
          nullable: true
          type: string
          format: uuid
        url:
          nullable: true
          type: string
          maxLength: 500
        new_window:
          type: boolean
        sequence:
          type: integer
          format: int32
        visibility:
          type: string
        required_member_roles:
          type: array
        is_mega_menu:
          type: boolean
        metadata:
          type: object
...
    CreateMenuRequest:
      type: object
      required:
        - website_id
        - name
        - new_window
        - sequence
        - visibility
        - required_member_roles
        - is_mega_menu
        - metadata
      properties:
        website_id:
          type: string
          format: uuid
        parent_id:
          nullable: true
          type: string
          format: uuid
        name:
          type: string
          maxLength: 120
        page_id:
          nullable: true
          type: string
          format: uuid
        url:
          nullable: true
          type: string
          maxLength: 500
        new_window:
          type: boolean
        sequence:
          type: integer
          format: int32
        visibility:
          type: string
        required_member_roles:
          type: array
        is_mega_menu:
          type: boolean
        metadata:
          type: object
...
    MenuList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Menu'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    PageBlock:
      type: object
      required:
        - id
        - page_id
        - kind
        - position
        - payload
      properties:
        id:
          type: string
          format: uuid
        page_id:
          type: string
          format: uuid
        kind:
          type: string
        position:
          type: integer
          format: int32
        payload:
          type: object
...
    CreatePageBlockRequest:
      type: object
      required:
        - page_id
        - kind
        - position
        - payload
      properties:
        page_id:
          type: string
          format: uuid
        kind:
          type: string
        position:
          type: integer
          format: int32
        payload:
          type: object
...
    PageBlockList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/PageBlock'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []